+++ CloudiQ/functions/Invoke-CloudiQApiRequest.ps1
@@ -1,19 +1,20 @@
 function Invoke-CloudiQApiRequest {
     [CmdletBinding()]
     param (
-        [Parameter(Mandatory = $True, Position = 0)]
+        [Parameter(Mandatory = $true, Position = 0)]
         [string]
         $Uri,
         [Parameter(Position = 1)]
         [string]
         $Method = "GET",
-        [Parameter(Position = 2, ValueFromPipeline = $true)]
+        [Parameter(Position = 2)]
         $Body,
         [Parameter(Position = 3)]
         [switch]
         $AsJson
     )
     try {
+        Write-Verbose -Message "Calling Cloud-iQ API"
         $restSplat = @{
             Uri         = "https://apiv1.crayon.com/api/v1/$Uri"
             Method      = $Method
@@ -23,7 +24,8 @@ function Invoke-CloudiQApiRequest {
                 'Authorization' = "Bearer $CloudIqAccessToken"
             }
         }
-        # If statement to include $Body, due to limitations in Invoke-RestMethod on Windows PowerShell.
+        # If statement to include $Body.
+        # Due to limitations in Invoke-RestMethod on Windows PowerShell, we have to define the JSON depth.
         if ($Body) {
             $result = Invoke-RestMethod @restSplat -Body ($Body | ConvertTo-Json -Depth 10) -ErrorAction Stop
         }
@@ -36,10 +38,13 @@ function Invoke-CloudiQApiRequest {
         break
     }
 
-    if ($AsJson) {
-        $result | ConvertTo-Json
-    }
-    else {
-        $result
+    Write-Verbose -Message "API results returned, either as object or JSON."
+    switch ($AsJson) {
+        $true {
+            $result | ConvertTo-Json
+        }
+        Default {
+            $result
+        }
     }
 }