+++ CloudiQ/functions/subscription/Get-CloudiQSubscription.ps1
@@ -21,6 +21,9 @@ function Get-CloudiQSubscription {
     .PARAMETER PublisherSubscriptionId
     The subscription Id from the portal.
 
+    .PARAMETER Detailed
+    Shows a more detailed view than the default one.
+
     .INPUTS
     Can either use the parameters Name or OrganizationId, or pipe any number of OrganizationId to the cmdlet.
 
@@ -68,6 +71,7 @@ function Get-CloudiQSubscription {
 
     # Depending on how we want to access subscriptions, Invoke-CloudiQApiRequest appropriatly
     if ($OrganizationName) {
+        Write-Debug -Message "Parameter OrganizationName used to get subscriptions."
         try {
             $OrganizationId = Get-CloudiQOrganization -Name $OrganizationName -ErrorAction Stop | Select-object -ExpandProperty Id
         }
@@ -78,9 +82,11 @@ function Get-CloudiQSubscription {
         $APICall = Invoke-CloudiQApiRequest -Uri "subscriptions/?organizationID=$OrganizationId" | Select-Object -ExpandProperty Items
     }
     elseif ($SubscriptionId) {
+        Write-Debug -Message "Parameter SubscriptionId used to get a specific subscription."
         $APICall = Invoke-CloudiQApiRequest -Uri ("subscriptions/" + $SubscriptionId)
     }
     else {
+        Write-Debug -Message "No parameter used to filter results, show all subscriptions available."
         $APICall = Invoke-CloudiQApiRequest -Uri "subscriptions" | Select-Object -ExpandProperty Items
     }
 
@@ -110,6 +116,8 @@ function Get-CloudiQSubscription {
     }
 
     if ($Name) {
+        Write-Verbose -Message "If the Name parameter is used, filter the results based on that name."
+        Write-Debug -Message "Value in Name parameter is set to $Name"
         $result = $result | Where-Object -Property ProductName -like $Name
         # Send warning if there are no results
         if (!$result) {
@@ -123,6 +131,6 @@ function Get-CloudiQSubscription {
             Write-Error ("No subscriptions found with that Subscription Id. Are you sure you meant " + $PublisherSubscriptionId + "?")
         }
     }
-
-    return $result | Sort-Object -Property 'Product'
+    Write-Verbose -Message "Returning results, sorted by product name."
+    return $result | Sort-Object -Property 'ProductName'
 }